Added functionality that saves linefit.sty with the latex export and fixed quadratic fit so that it does not allow for fixed variables. Need to test the linefit.sty on a Linux based system though

+++ LineFit WS/LineFit/src/linefit/GraphOptionsMenu.java
@@ -942,7 +942,7 @@ class GraphOptionsMenu extends JFrame
 			
 		SystemIOHandler.LaTexGraphWidthInCm = Double.parseDouble(LaTexWidthField.getText());
 		SystemIOHandler.LaTexGraphHeightInCm = Double.parseDouble(LaTexHeightField.getText());	
-		SystemIOHandler.exportFontSize = (float) (double) exportFontSize.getValue();
+		SystemIOHandler.exportFontSize = (Float) exportFontSize.getValue();
 		
 		graphingArea.resultsAreDisplayedOnGraph = displayResultsOnGraphCheckBox.isSelected();
